Add files via upload
Pixel Gun 3D Methods and ClassesRespository by HorridModz (User123456789#6424).
This resource is PUBLIC. You can DISTRIBUTE it to ANYONE with credit.

This respository took tens and tens of hours to make. It is the combined work of searching for every single possible string I could think of and writing down the results.
I am sorry it is so messy. It was meant to be for personal notes, but it became so big! Not everything will do what it's caption says. Most things are untested.

This repository includes:
-Labeled methods, classes, fields, and notes. Most of these are from pixel gun 3d version 16.6.1, 22.3.2, 22.4.0, or 22.4.3.
-Notes about how to find things and deobfuscate them
-Script templates

How to use this repository:
-This is not a tutorial or a mod. It is just a document of my personal notes. This is a resource for hex patching or creating mod menus. If you do not know what you are doing, do not use this.
-Search for a keyword such as "silent aim" or "unlock all"
-Choose one of the methods (or fields or whatever) to use
-Try to find this in the current game version
-Make a mod menu or gameguardian script using the offset
-Test it and see what it does. Feel free to change the caption if it does not do what it said it does. Most don't, as they are untested.

I hope you enjoy using this resource. I can make repositories like this for any game. (Of course,as long as you have an unobfuscated and unprotected version of the game. I'm not going to give you a bunch of free fire hacks xD I'm not that smart.) Tell me if you would like a respository for a certain game.